Update issue 1450
Chaning the exception hierarchy to explicistly test if the directory has been created.  This was a problem in windows that would throw other kinds of Windows specific errors, not just an OSError.


Former-commit-id: 8e3105abd04c93edbbce8d054d022716b53fb7fb

+++ invest_natcap/nutrient/nutrient_biophysical.py
@@ -21,11 +21,10 @@ logging.basicConfig(format='%(asctime)s %(name)-15s %(levelname)-8s \
 class InvalidNutrient(Exception):pass
 
 def make_folder(folder):
-    try:
-        os.makedirs(folder)
+    if not os.path.exists(folder):
         LOGGER.debug('Making the folder %s', folder)
-    except OSError:
-        # Thrown when folder already exists
+        os.makedirs(folder)
+    else:
         LOGGER.debug('Folder %s already exists, deleting and recreating', folder)
         shutil.rmtree(folder)
         os.makedirs(folder)
